chore(`ffi`): fix compilation errors
